Move matplotlib tests to matplotlib-pyodide
