+++ packages/matplotlib-pyodide/test_data/canvas-custom-font-text-chrome.png
