+++ packages/matplotlib-pyodide/test_data/canvas-custom-font-text-firefox.png
