+++ packages/matplotlib-pyodide/test_data/canvas-firefox.png
