+++ packages/matplotlib-pyodide/test_data/canvas-image-affine-chrome.png
