+++ packages/matplotlib-pyodide/test_data/canvas-image-affine-firefox.png
