+++ packages/matplotlib-pyodide/test_data/canvas-image-affine-safari.png
