+++ packages/matplotlib-pyodide/test_data/canvas-image-chrome.png
