+++ packages/matplotlib-pyodide/test_data/canvas-image-firefox.png
