+++ packages/matplotlib-pyodide/test_data/canvas-image-safari.png
