+++ packages/matplotlib-pyodide/test_data/canvas-math-text-chrome.png
