+++ packages/matplotlib-pyodide/test_data/canvas-math-text-firefox.png
