+++ packages/matplotlib-pyodide/test_data/canvas-math-text-safari.png
