+++ packages/matplotlib-pyodide/test_data/canvas-polar-zoom-chrome.png
