+++ packages/matplotlib-pyodide/test_data/canvas-polar-zoom-firefox.png
