+++ packages/matplotlib-pyodide/test_data/canvas-polar-zoom-safari.png
