+++ packages/matplotlib-pyodide/test_data/canvas-safari.png
