+++ packages/matplotlib-pyodide/test_data/canvas-text-rotated-safari.png
