+++ packages/matplotlib-pyodide/test_data/canvas-transparency-chrome.png
