+++ packages/matplotlib-pyodide/test_data/canvas-transparency-firefox.png
