+++ packages/matplotlib-pyodide/test_data/canvas-transparency-safari.png
