+++ packages/matplotlib/extras/setup.cfg
@@ -1,5 +1,2 @@
-[rc_options]
-backend=module://matplotlib_pyodide.wasm_backend
-
 [libs]
 system_freetype = true
